Fix quest filters for necros and druids
All necros share the same type_id = 99, same with druids (100) and
catapults (101). Let's filter only "kill n mobs" quests with the weird
mob identifier based on face and `type_id`, and all in other cases we
have a pointer to a unit, so we can use a more stable `server_id`.